Completely hide the label of the search input
This commit makes the paragraph that is the label of the search input
completely hidden by replacing the `class="sr-only"` with `hidden`
attribute.

The reason is because the fact that screenreader can still access the
hidden element as long as the `input` is labeled with `aria-labelledby`.

Reference: https://www.sarasoueidan.com/blog/accessible-icon-buttons/#technique-%232%3A-accessible-visually-hidden-text-with-hidden-and-aria-labelledby