Remove the "Skip to main content" link
This commit removes the "Skip to main content" link. It's because the
amount of the navigable items before the `main` element  are only two. As a result, the users can
navigate to the main content quickly. So, having that link may not
be necessary at all.

Reference: https://webaim.org/techniques/skipnav/#multiple